ar71xx: fix WNDR3x00 factory image generation
Factory images for WNDR3700v2 and WNDR3800 should contain these model IDs
for the device: field in their DNI tags, not WNDR3700.

This regressed in r29434, which redid WNDR3x00 model detection. The sixth
parameter to Image/Build/Netgear is only used as the -B argument to
mkdniimg, which is used to set the device: field. In r29434, this was
erroneously changed to be WNDR3700 for all models. The tools to flash
factory images (U-Boot's TFTP server and the factory software's upgrade
utility) may refuse to honor images with incorrect device: fields in their
DNI tags.

Signed-off-by: Mark Mentovai <mark@moxienet.com>